(최종)Doc_agent_project/Doc_agent_project/backend/requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.12

# Database and Vector Store
supabase==2.7.4
pgvector==0.3.5

# Embeddings and ML
sentence-transformers==3.3.1
torch==2.5.1
numpy==1.26.4

# LLM Integration
openai==1.54.5
httpx==0.27.2

# Document Processing
pdfplumber==0.11.4
python-docx==1.1.2
pypdf==5.1.0

# Text Processing
langchain==0.3.7
langchain-text-splitters==0.3.2

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Development
black==24.10.0
flake8==7.1.1
pytest==8.3.3
